pyproject.toml
[tool.poetry]
name = "cdk-pipelines-demo"
version = "0.1.0"
description = "CDK Pipelines Demo"
authors = ["Justin Stewart <justin@justinstewart.dev>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
aws-cdk-lib = "^2.158.0"

[tool.poetry.group.dev.dependencies]
yawsso = "^1.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
